Move Github OAuth to a library helper function
