+++ recipes/ibmz_ci.rb
@@ -18,6 +18,7 @@
 
 class ::Chef::Recipe
   include OSLDocker::Helper
+  include OSLGithubOauth::Helper
 end
 
 node.default['osl-jenkins']['secrets_item'] = 'ibmz_ci'
@@ -48,6 +49,7 @@ node.default['osl-jenkins']['restart_plugins'] = %w(
   ssh-slaves:1.26
   token-macro:2.4
   durable-task:1.17
+  docker-commons:1.11
   docker-java-api:3.0.14
   docker-plugin:1.1.3
   plain-credentials:1.4
@@ -88,7 +90,6 @@ node.default['osl-jenkins']['restart_plugins'] = %w(
 )
 
 node.default['osl-jenkins']['plugins'] = %w(
-  docker-commons:1.11
   pipeline-model-extensions:1.1.3
   emailext-template:1.0
   pipeline-stage-tags-metadata:1.1.3
@@ -162,119 +163,21 @@ docker_cloud =
     docker_cert['cert'],
     docker_cert['chain']
   )
+
+github_oauth =
+  add_github_oauth(
+    client_id,
+    client_secret,
+    admin_users,
+    normal_users
+  )
+
 jenkins_script 'Add Docker Cloud' do
   command docker_cloud
 end
 
 jenkins_script 'Add GitHub OAuth config' do
-  command <<-EOH.gsub(/^ {4}/, '')
-		import hudson.security.*
-		import jenkins.model.*
-		import org.jenkinsci.plugins.GithubAuthorizationStrategy
-		import hudson.security.AuthorizationStrategy
-		import hudson.security.SecurityRealm
-		import org.jenkinsci.plugins.GithubSecurityRealm
-		import hudson.security.ProjectMatrixAuthorizationStrategy
-		import hudson.security.csrf.DefaultCrumbIssuer
-
-		Jenkins.instance.crumbIssuer = new DefaultCrumbIssuer(true)
-
-		// Authentication
-		String githubWebUri = 'https://github.com'
-		String githubApiUri = 'https://api.github.com'
-		String clientID = '#{client_id}'
-		String clientSecret = '#{client_secret}'
-		String oauthScopes = 'read:org'
-		SecurityRealm github_realm = new GithubSecurityRealm(githubWebUri, githubApiUri, clientID, clientSecret, oauthScopes)
-		//check for equality, no need to modify the runtime if no settings changed
-		if(!github_realm.equals(Jenkins.instance.getSecurityRealm())) {
-				Jenkins.instance.setSecurityRealm(github_realm)
-				Jenkins.instance.save()
-		}
-
-		// Authorization
-		class BuildPermission {
-			static buildNewAccessList(userOrGroup, permissions) {
-				def newPermissionsMap = [:]
-				permissions.each {
-					newPermissionsMap.put(Permission.fromId(it), userOrGroup)
-				}
-				return newPermissionsMap
-			}
-		}
-
-		auth_strategy = new hudson.security.ProjectMatrixAuthorizationStrategy()
-
-		authenticatedPermissions = [ "hudson.model.Hudson.Read" ]
-		authenticated = BuildPermission.buildNewAccessList("authenticated", authenticatedPermissions)
-		authenticated.each { p, u -> auth_strategy.add(p, u) }
-
-    anonPermissions = [ "hudson.model.Hudson.Read" ]
-		anon = BuildPermission.buildNewAccessList("anonymous", anonPermissions)
-		anon.each { p, u -> auth_strategy.add(p, u) }
-
-
-		adminPermissions = [
-      "com.cloudbees.plugins.credentials.CredentialsProvider.Create",
-      "com.cloudbees.plugins.credentials.CredentialsProvider.Delete",
-      "com.cloudbees.plugins.credentials.CredentialsProvider.ManageDomains",
-      "com.cloudbees.plugins.credentials.CredentialsProvider.Update",
-      "com.cloudbees.plugins.credentials.CredentialsProvider.View",
-      "hudson.model.Computer.Build",
-      "hudson.model.Computer.Configure",
-      "hudson.model.Computer.Connect",
-      "hudson.model.Computer.Create",
-      "hudson.model.Computer.Delete",
-      "hudson.model.Computer.Disconnect",
-      "hudson.model.Computer.Provision",
-      "hudson.model.Hudson.Administer",
-      "hudson.model.Hudson.Read",
-      "hudson.model.Item.Build",
-      "hudson.model.Item.Cancel",
-      "hudson.model.Item.Configure",
-      "hudson.model.Item.Create",
-      "hudson.model.Item.Delete",
-      "hudson.model.Item.Discover",
-      "hudson.model.Item.Move",
-      "hudson.model.Item.Read",
-      "hudson.model.Item.ViewStatus",
-      "hudson.model.Item.Workspace",
-      "hudson.model.Run.Delete",
-      "hudson.model.Run.Replay",
-      "hudson.model.Run.Update",
-      "hudson.model.View.Configure",
-      "hudson.model.View.Create",
-      "hudson.model.View.Delete",
-      "hudson.model.View.Read",
-      "hudson.scm.SCM.Tag"
-		]
-    #{admin_users}.each { au -> user = BuildPermission.buildNewAccessList(au, adminPermissions)
-      user.each { p, u -> auth_strategy.add(p, u) }
-    }
-
-		userPermissions = [
-			"hudson.model.Item.Build",
-			"hudson.model.Item.Cancel",
-			"hudson.model.Item.Configure",
-			"hudson.model.Item.Create",
-			"hudson.model.Item.Delete",
-			"hudson.model.Item.Discover",
-			"hudson.model.Item.Read",
-			"hudson.model.Item.ViewStatus",
-			"hudson.model.Item.Workspace"
-		]
-    #{normal_users}.each { nu -> user = BuildPermission.buildNewAccessList(nu, userPermissions)
-      user.each { p, u -> auth_strategy.add(p, u) }
-    }
-
-
-		//check for equality, no need to modify the runtime if no settings changed
-		if(!auth_strategy.equals(Jenkins.instance.getAuthorizationStrategy())) {
-				Jenkins.instance.setAuthorizationStrategy(auth_strategy)
-				Jenkins.instance.save()
-		}
-
-	EOH
+  command github_oauth
 end
 
 ruby_block 'Set jenkins username/password if needed' do
